sdk testing cleanup
* update packages
* superToken cfa interface cleanup
* yarn.lock update
