+++ packages/sdk-core/package.json
@@ -63,8 +63,8 @@
         "@graphql-codegen/typescript": "^2.4.3",
         "@graphql-codegen/typescript-operations": "^2.2.4",
         "@graphql-typed-document-node/core": "^3.1.1",
-        "@nomiclabs/hardhat-ethers": "2.0.3",
-        "@nomiclabs/hardhat-waffle": "^2.0.2",
+        "@nomiclabs/hardhat-ethers": "2.0.5",
+        "@nomiclabs/hardhat-waffle": "^2.0.3",
         "@nomiclabs/hardhat-web3": "^2.0.0",
         "@typechain/ethers-v5": "^8.0.5",
         "@typechain/hardhat": "^3.1.0",
@@ -83,17 +83,17 @@
         "eslint-plugin-eslint-comments": "^3.2.0",
         "eslint-plugin-import": "^2.25.4",
         "eslint-plugin-prettier": "^3.3.1",
-        "ethereum-waffle": "^3.4.0",
+        "ethereum-waffle": "^3.4.4",
         "get-graphql-schema": "^2.1.2",
-        "hardhat": "^2.8.3",
-        "mocha": "^9.2.0",
+        "hardhat": "^2.9.1",
+        "mocha": "^9.2.2",
         "npm-run-all": "^4.1.5",
         "nyc": "^15.1.0",
-        "ts-node": "^10.4.0",
+        "ts-node": "^10.7.0",
         "typechain": "^6.1.0",
         "typedoc": "^0.22.11",
-        "typescript": "^4.5.5",
-        "web3": "^1.7.0"
+        "typescript": "^4.6.2",
+        "web3": "^1.7.1"
     },
     "resolutions": {
         "graphql": "^16.0.0"
